Abort service stop action if the parent proc cwd is held by a resource
This patch avoids killing the service stop at the fs stop step
because of "fuser -k". In this situation, the service is left in
an intermediate state and the shell that triggered the action is
killed. This is hardly the expected result of a stop action.

So better to abort the action early, stating that the user should
relocate its shell before action.