Remove the prevg and postvg tag heuristics
Now only rely on the rid and subsets sorting logic to let users
define in which order to start and stop the resources.

This heuristics were used only by the drbd and md resources.

Beware drbd users, now the postvg tagged resources are considered
prevg, because 'drbd' is alphabetically before 'vg'.