Set the zone tag exclusion selector for all disk resources
This exclusion selector was used only for zpool resources, forcing
it to be actioned outside the main disk action.

Setting it for all disk resources allows to merged the zpool
action in the main disk action and allows to deffer disk
actions other than zpool after the zone ready state if desired.