Add a docker ip resource driver
So that we can benefit from the abort_start() safety net and good
resource health checking.

We usually used pipework in a container post_start trigger, but this
practice had those two shortcomings.

The syntax of such a resource is :

[ip#0]
ipdev = lxcbr0
ipname = 10.0.3.3
netmask = 255.255.255.0
gateway = 10.0.3.1
container_rid = container#0
tags = docker

(iow, quite close to the syntax of a zone shared ip)

For now we still expected to find a patched version of pipework
in /opt/opensvc/bin/, but we will reimplement natively what we
need in a following patch to remove this dependency.