Detect and remove resource device holders on stop
Stopping a resource serving a device can succeed only if the
device has no holders left.

Holders can be handled by another service resource. In this case
error out, because either the service configuration is incorrect
with regard to resource ordering in their subset, or the user
as asked for a specific resource stop not realizing other resources
depend on it.

But if all holders are not handled by the service, try agressively
to remove them.

This strategy applies to
* linux fs, because this os can layer loopback devices over files of
the fs
* linux vg, because this os can layer devmaps over lvs of the vg

The 1st case happens with docker lvm storage backend, which
create loops devices in the docker data dir

The 2nd case happens with docker lvm-direct storage backend, which
layers thinp volumes over the lvs the service hands to dockerd.

This generalized approach deprecates the "kpartx -d ..." calls, as
the partitions created through kaprtx are found as unhandled
holders, and removed through "dmsetup remove ..."