Allow user to set a rcmd parameter in lxc container resources
This rcmd parameter allows users to override the command chosen
by the agent to submit commands in the container.

It can be used to enable the --elevated-privileges for example:

[container#0]
type = lxc
name = foo
rcmd = lxc-attach -e -n foo --

This parameter might be extented to other containers later, as
it might be relevant to let users chose a custom ssh key to enter
a vm.