Check that the lxc container config file is installed in the sys dir
Usually this is /var/lib/lxc/<container>/config

Since systemd can be embedded in the containers, the config file
should contain definition of the kill signals to use to shutdown
the container. Thus the config file must be available to the
lxc-stop command. But lxc-stop still lack a -f flag opensvc could
use to point our private config file. lxc-stop only looks for the
config file in its buildin defined location.

This patch adds a resource status check for the presence of the
config file in the standard location, and adds it to the resource-
specific files to sync with sync#i0.