Fix "monitor_action = freezestop" corner case
Considering "intf" is the interface used to communicate with the
dns, if the following conditions are met:

1/ the service is setup for HA with OpenHA
2/ the monitor action is freezestop
3/ an interface or a container is monitored
4/ "intf" goes down
5/ one of the ip resource ipname or one of the container name is
resolvable only through the dns (say "myname")
6/ no dns caching daemon is running

... then the freezestop monitor action is triggered and the
service goes down, but fails because "myname" can't be resolved
to an ip. The OpenHA service thus transitions to STOP_FAILED
and the backup node will not take over.

This patch introduces a privately held getaddrinfo cache.
The cache is used only on the stop action code path, thus
start and status still report dns error as soon as the system
reports them.

Misc policies:
- The cache expires at 16 minutes.
- The entries are stored in /opt/opensvc/var/cache/addrinfo/
- Every successful getaddrinfo refreshes the cache.