Support node.prkey node.conf parameter
This parameter can be used to force a prkey instead of relying
on the agent internal algo, which on certain situation can not
garanty stability.

This happens on Linux, where the uuid.getnode() function bases
the prkey on a mac address ... but hypervisors tend to be to
dynamic in their ip link list for that approach to be reliable.