Linux VG driver fixes
1/
Fallback to doing a pvscan if systemd fails to do it itself.
The VG tagging need the vgs command, which fails if the pvscan
has not run since the VG PVs activation. Systemd is normally
responsible for this, but sometimes it fails to do it.

2/
Make sure the start adds the tag if they are not set, even if
the VG is already up.