Fix not started apps declared in sysv mode
When launcher script are hosted in a fs not mounted when the
start action is launched, the apps would not be added to the
app default resource set after the fs mount.

The code to do so was present, but the new resources were added
in the Svc context, and we needed them in the RsetApp context.

This patch, just points back the default RsetApp.resources to
the list stored in Svc::type2resSets["app"].resources after the
last minute sysv re-inventory.