Linux fs resource stop enhancements
When the devices supporting a Linux fs report i/o error, the
mount point existance test would report an error and the resource
stop was aborted.

Test if the os.path.exists() OSError errno is 5, indicating we
are in this situation, and don't abort the stop.

In this situation, the fuser will not be able to determine which
process hold the resource. So implement a build-in routine
searching in /proc/*/fd/* + /proc/*/cwd + /proc/*/exe for process
pinning the mount point, and kill them with SIGKILL.