Do not TOC because of monitored non-up resources in frozen encap services
It is a common use-case for users to freeze only the encap part of
HA services. In this case, we don't want to trigger a service
failover through the resource_monitor because of encap monitored
resources which might legitimately go down because of the
advertized maintenance.

The encap status json structure already contains the frozen
status of the encap part, so we just have to shortcut the
resource_monitor in the master service for frozen encap parts.