Don't TOC for monitored resources in n/a state
App resource can degrade to n/a if the aggregate status of
reference ressources is not up (ip, fs, disk and containers).

If the app is monitored, this would trigger a node suicide (TOC).
This is clearly not the expected behaviour.

This patch makes sure a n/a resource will not trigger a TOC.