Allow to set containerization capping settings in resources
Not just in the DEFAULT service section.

The intent is to create the following containerization hierachy
for a service:

├─testmd
│ ├─app:ga2
│ │ ├─app.4
│ │ │ └─31621 dd if=/dev/urandom of=/dev/null bs=1M count=1000
│ │ └─app.3
│ │   └─31627 dd if=/dev/urandom of=/dev/null bs=1M count=1000
│ └─app:ga1
│   ├─app.2
│   │ └─31542 dd if=/dev/urandom of=/dev/null bs=1M count=1000
│   └─app.1
│     └─31546 dd if=/dev/urandom of=/dev/null bs=1M count=1000

Where testmd is the service, app:* are subsets and app.* are
resources.

This patch alos add more properties to a new subset upon addition
to the service:
- a "svc" backpointer to the service
- a "rid" to identify the subset
- the containerization capping settings

Finally fix a type in the container_cpu_shares parameter in the
service dictionnary (missing trailing 's').