Exit early from resource status eval when pg or parent pg is frozen
If the resource process group or one of its parent pg is frozen,
chances are that the normal resource evaluation gets stuck.

It is the case for lxc resource evaluator at least.

In this case, just degrade the status to WARN and log the reason.

This patch also fixes the "pg freeze" and "pg thaw" --rid and
--tags honoring.