Allow users to set a python wrapper script
Some systems having multiple libraries versions installed might
want to use a wrapper script to select which libraries to link.

Such a wrapper installed in /opt/opensvc/bin/python would be
removed upon each bin/postinstall run (hence each agent upgrade).

To avoid this issue, this patch standardize the bin/python.wrapper
name for this wrapper, and if present it is chosen as the target
of the bin/python symlink.