Solaris zone driver enhancement
fs and ip resources have special behaviour when linked to a zone
using the "zone = <zonename>" resource parameter. This
configuration requires that zone referenced zone is handled by
a declared resource in the same service.

Users might want to disable the container resource actions using
the "noaction" tag. In this case, the noaction behaviour
propagates to the ip and fs resources referencing the container.

This is not the expected behviour. This patch makes sure those
resource are still actioned even if their referenced zone has
the noaction tag.