Do not create cgroup infra structure upon resources status evaluation
Resource status eval call the rcContainer frozen method, which in
turn calls a utility function to get the cgroup path on Linux. In
case the cgroup path does not exists, it gets created by this
utility function.

This patch adds a "create" boolean parameter to this function, so
that the frozen() caller can set it the False.