Fix group status aggregation
Now that the "disk" resource set type as different subtypes
(disk.scsireserv, disk.md, disk.vg, ...), the aggregated status
for group got reseted for each subtype.

As a consequence, if a resource from the disk.scsireserv is "down"
and all other resources are "up", include a disk.lvm resource, the
"disk" group status was unduly evaluated as "up".

This patch takes into account that the rset_status[rs.type] might
already exist when looping the rs.type subtype, and in this case
merges the status instead of overloading.