Make sure we never create a DEFAULT#0 resource in a service config
The create and update svcmgr actions would create a DEFAULT#<n>
resource section when parsing a --resource '{"rtype": "DEFAULT",
...}' parameter.

To populate the DEFAULT section, the agent expected no rtype
in the -ressource dictionnary. Finding a rtype, it thought it to
be a normal resource, and thus allocated a rid increment-suffix.

This patch make sure we tolerate "rtype": "DEFAULT" too.