Add the "sync restore" action to Amazon s3 sync driver
This action restores the last full and all incremental backups.

Beware, this is primarily meant to restore staged services, ie
services backed up to a bucket and removed from nodes. In this
case the service provisioning plus this sync restore action
provide a simple procedure to restore the service.

If you sync restore on a live service, you may harm your data.