Fix apps and networked fs encap resource start in systemd containers
systemd on-demand loading will let us start the encap service before
the network is fully initialized, causing start issues with nfs mounts
and listening apps.

This patch adds to the lxc container's operational() function a
test for systemctl is-active default.target, which delays the
encap start to after the container is fully functional.