Stop trying to shutdown remote vm on service stop
If the container resource was not in DOWN state, the shutdown
was triggered. But a vm running on a remote should never be
shutdown on service stop on a node not running the vm. The
WARN container state could lead to an invalid shutdown.

This patch uses the up() container method instead of the status().
up() is alwoys False if the vm is not local, which suits the
need.