Support docker ip stop on rhel 7.2
In rhel 7.2, the veth added to the docker container is magically
suffixed with @if<n>. All ip commands must be passed to the
unsuffixed interface, which is the case with the resource start
ip commands.

The resource stop ip commands are passed to the interface
actually detected in the container namespace, which is the
suffixed interface name.

This patch detects the presence of the suffix and strips it.