Sysreport enhancement
Support the case where a collected path changes from dir to regular
file. In this case, the code failed to remove the archived dir and
the new file ended up as a child node of the directory.

As a consequence, a change on this file was reported and commited
server-side upon each sysreport execution. Flooding the timeline
with misleading information.