Honor the group_status() fine-grained group exclusion
The group_status() function takes a excluded_groups list parameter.
Entries of this list can be in the form <group>.<rtype>, like
"disk.drbd".

In this case, group_status() still aggregated the status of
the excluded resources because it used the resourceset (disk)
status.

This patch makes

1/ the group_status() loop over resources after
filtering out the excluded resources to aggregate only the
expected status.

2/ uses this new capability to filter out disk.drbd from the
group_status() call to decide if we need to run the launchers
status evaluation.