Support deleting the default network route within the docker containers
A docker ip resource can now set the del_net_route and network
parameters to instruct the driver it must delete the default
net route.

This scenario is interesting with hosts attaching the docker
containers to backend bridges. Containers having their ip in
a cloud private subnet, all ips being routed at the cloud level
to a specific cloud instance. Deleting the default net route
allows all packets to be sent out of the container for the
host to take the routing decisions: return to the local bridge if
the destination container is local or out to the cloud routing
if the destination container is hosted on another host.

This setup works well on Google Compute Engine instances.