Add the DEFAULT.show_disabled service config parameter
Switch back the default "show_disabled" to true, to maintain to
legacy behaviour.

Introduce a service default section parameter "show_disabled" to
specify the service default behaviour with regards to disabled
resources inclusion in the print status output.

The command line parameter take precedence over the service
parameter.

Also add a --hide-disabled command line parameter, in addition
to the --show-disabled, to force hiding when the service default
is set to true.