Add the pull svcmgr action
The pull action fetch the configuration of a service from the
collector, writes it in etc/<svcname>.env and installs the
executable symlink.

The intended usage is to create a service collector-side, then
pull it on nodes to provision and start it.

Example:

$ sudo svcmgr -s svc1 pull
INFO    clementine /opt/opensvc/etc/svc1.env pulled