Add substitution of the %images% and %instances% words in docker wrapped cmds
The docker command wrap the docker client command, setting
automatically the socket parameter to join the service-private docker
daemon.

The %images% and %instances% words in the wrapped command are now
replaced by, respectively, the set of docker instance names and
images for container resources passing the --tags, --rid and
--subsets filters.

This is useful to remove all instances of a service or all instances
of resources with a tag like "frontend".

Note the opensvc filters must be positioned before the docker command
in the arguments list.

Example:

$ sudo svc1 --tags frontend docker rm %instances%