Preserve service configuration comments and ordering on set/unset
The set and unset actions used the ConfigFile parser to edit the
service configuration. Upon config write, ConfigFile cannot
maintain the initial ordering, and comments (which are ignored
by the parser) are dropped.

This patch implements the set and unset operations using direct
file read => buffered config mangle => file write, which allow
us to maintain ordering and comments.

If a set parameter is not found in an existing section, it gets
added at the end of the section.

If a set section is not found, it gets added at the end of the
configuration file.

Multiline values are concatenated before comparison with the
set action target value.