Scsi reservation driver enhancements
1/

Detect sg_persist "error opening file" and don't enter the
Unit Attention wait loop in this case.

This change speeds-up the error reporting.

2/

Implement a fallback to run sg_persist on the multipath when
sg_persist fails on a path.

For whatever reason, the sg_persist doesn't error out on the
multipath, even if it's read-only too.

This is not a perfect solution, as we don't control which path
the command is routed too.