Linux Raw disk driver fix
Don't split present/not-present device paths in separate lists
in the driver __init__().

Upon service start, raw devices with devs served by a vg handled
by the service are not yet present. If we sort out those devices
at __init__() time, they won't ever be activated.

This patch move the devs splitting at the head of each action
entrypoint in the driver.

With this change, raw devices on lv in a vg handled by the service
works as expected.