zfs fs resource fix
1/

Make sure we set the dataset zoned property to on only if the
zone resource parameter is set or the resource is tagged 'encap'

2/

Make sure we set the dataset mountpoint property to the mnt
resource parameter whenever it is set to a different value. The
mnt parameter value is prefixed with the zone path only if the
zone resource parameter is set.

With this patch a fs can be in any zoned/mountpoint initial state
and the resource can be zoned, global or encap.