Implement a dummy vm_hostname() docker driver method
As the docker driver is a container class child, it inherits its
parent class vm_hostname() method.

Though this method is not adapted nor necessary for docker, as
it tries to ssh the container. We could define a docker exec
method to fetch the instance hostname if it was useful.

Instead, this patch defines a dummy vm_hostname method in the docker
driver class.