Fix the node.conf defaults fetching for options in absent sections
In the case of the host_mode parameter, a default is set to TST.
This default is correctly reported to 'nodemgr get --param node.
host_mode' if the [node] section exists, but fails if the section
does exist.

This patch detects the abence of the section and, instead of
reporting that as an error, create the section in the in-memory
parsed config. Also trap the NoOptionError for options with
no default value, to exit with an !0 retcode and display a
proper error message.

With this patch:

$ file etc/node.conf
etc/node.conf: cannot open `etc/node.conf' (No such file or directory)

$ sudo nodemgr get --param node.host_mode
TST

$ sudo nodemgr get --param node.foo
option 'foo' not found in section 'node'