Fix infinite loop with zpool + scsireserv + missing disklist cache
Upon scsireserv resource status evaluation, the resource gets
the disklist from the zpool co-resource, which (if the disklist
cache is not present) evaluates the scsireserv resource status
to decide if it can reexport the disklist cache.

This patch replaces the group_status() call by a more localized
is_up() to decide if the disklist cache can be reexported.