Fix a Linux prstop warning
The prstop uses the co-resource disklist() after the co-resource
has been stopped. On linux the disklist starts with a devlist,
including the LV of a VG, and resolves this list to disks.

As the VG is stopped the LV are not accessible, and prstop
emits the "can not stat" warning.

This patch downgrades the warning to a debug.