Check if a zpool is suspended before trying a dataset write test
If all members of a zpool are unavailable and the pool failmode
is set to the "wait" default value, the write test hangs and thus
the resource monitoring hangs.

This patch makes sure even in this situation, the resource
monitoring does not hang by skipping the write test and logging
the suspended pool situation.

Example:

$ s11zone print status --refresh
s11zone
overall                   warn
'- avail                  warn
   |- fs#2           .... warn       mpool/dset1@/mpool/dset1
   |                                 # pool %s is suspended # fs is not
   |                                 writable
   |- fs#1           .... up         /tmp@/system/zones/testzone/root/mytmp
   '- container#0    .... up         testzone