Add the zfs and zpool compliance objects
And their common zprop parent class.

Those module are straighforward. The nice part is in the form
definition, which knows the candidates of most properties.

Example:

$ ./zfs.py info
    [
        {
            "op": "=",
            "name": "rpool/swap",
            "value": "discard",
            "prop": "aclmode"
        },
        {
            "op": "<",
            "name": "rpool/swap",
            "value": 1,
            "prop": "copies"
        },
        {
            "op": ">",
            "name": "rpool/swap",
            "value": 0,
            "prop": "copies"
        },
        {
            "op": "<=",
            "name": "rpool/swap",
            "value": 1,
            "prop": "copies"
        },
        {
            "op": ">=",
            "name": "rpool/swap",
            "value": 1,
            "prop": "copies"
        }
    ]

$ ./zfs.py test
property aclmode current value - is not = discard.
property copies current value 1 is not < 1.
property copies current value 1 is > 0. on target.
property copies current value 1 is <= 1. on target.
property copies current value 1 is >= 1. on target.