Fix stack in docker_running()
The symptom is:

  File "/opt/opensvc/lib/rcDocker.py", line 357, in docker_running
    exe = os.path.realpath(exe)
  File "/usr/lib64/python2.7/posixpath.py", line 367, in realpath
    path, ok = _joinrealpath('', filename, {})
  File "/usr/lib64/python2.7/posixpath.py", line 406, in _joinrealpath
    path, ok = _joinrealpath(path, os.readlink(newpath), seen)
OSError: [Errno 2] No such file or directory: '/proc/6483/exe'

Though we test the presence of /proc/<pid> just before, it can
happen the exe file is not there or the /proc/<pid> disappears
before the realpath() call.

Catch the realpath() OSError exception and consider the daemon is
not running in this case. And cleanup the pid file.