Support SunOS remapping of disks from the disklist cache
The master node sends the zfs pools disklist with its local device
paths as entries. But the device path can be different on slave
nodes (controller index change for example).

In this situation, the persistent reservation status eval fails on
the slave nodes.

This patch applies a validation and remapping of the disklist
entries, based on the presence of the wwid in the device path.