Really try to umount anyway when the message is printed
Patch 9bb487b680dc5bd46234065bd4593181d256779f was meant to try
umounting the fs even on i/o erroring fs, but failed to
actually do that even if the printed message says it does.

This patch simply removes the exception raising causing the
interruption of the intended workflow, which was a leftover of
the previous behviour.