Fix disklist() on zpools with emcpower devices
emcpower devices remaps the s[0-8] suffix to [a-g].

This patch thus remaps emcpower<n>[a-g] to emcpower<n>c instead
of remapping to emcpower<n>s2, which does not exists.

Incindentally, it also fixes the scsi reservation driver, which
uses the zpool disklist() to fetch the device list the driver
handles.