Deprecate the python wrapper usage
Make svcmon, svcmgr and nodemgr symlink to a new opensvc shell
script.

The shell script is in charge of the execution environment setup:
* which python binary to use, default: python
* LD_PRELOAD, LD_LIBRARY_PATH, default: none
* python additional args if specified (OSVC_PYTHON_ARGS), default: none

The system defaults are sourced in:
* /etc/defaults/opensvc
* /etc/default/opensvc
* /etc/sysconfig/opensvc

The old python svcmon, svcmgr and nodemgr source files are relocated
in lib/, suffixed with .py and no longer have a shebang.