Add the 'osvc_root_path' containers parameter
This 'osvc_root_path' parameter allow users to define where the
agent is installed in the container.

If left undefined, the container agent is assumed to be installed
in the LSB FSH.

With this patch, users can have a LSB installation in the
hypervisor and have some containers with a developper agent tree.