Fix the .stonith link behaviour
When the opensvc shell wrapper is called through a .stonith link,
the arguments should be reset.

OpenHA executes the .stonith link with the "start" argument. If
we don't reset $@, svcmgr ends up being called with the
"stonith_start" action, which is invalid.

Also log the invalid action to node.log, so non-interactive
executions leave a trace of such problems.