Support netfs provisioning as noop
Fallback to provFs when no provFs<fsType> is implemented, so that
we don't have to add and maintain dummy files for netfs
provisioners.

Make provFs behave correctly when no mkfs or info properties are
set, which is the case when the fallback to provFs is activated.
In this situation check_fs() will return ok, and as a consequence
no mkfs will be attempted.