Support the zpool resource "name" property
"name" is the standard property name for pooling disk resources.
Continue to support the deprecated "poolname" attribute for
compatibility.

This patch also adds the "deprecated" boolean property the the
svcDict Keyword class. A deprecated Keyword is not rejected by
the service config validator but it does not appear in the
interactive service creation forms nor in the templates.