Add the deprecated sections notion to the service configuration dict
Resources like zpool can be declared as pool#n, disk#n+type=pool
or disk#n+type=zpool. The later is preferred, the 2 former forms
are deprecated but are still supported for compatibity.

This patch blocks the generation of the template.<driver>.<type>.env
in the docs for <driver> and <driver>.<type> listed as deprecated.

Thus we no longer generate the template.disk.pool.env template.