Beautify the build validation messages
Don't display build errors for services not explicitely selected.
=> svcmon do not report build errors

Don't display the svcmgr usage message if a service selection
was set but trimmed because of unvalidated services in the
selection

Replace all log from svcBuilder.build() by raising the message
in an ex.excInitError. The caller is now in charge of doing what
he wants with the error messages.