Don't stack trying to parse encap svc config file mtime
when the container is down, fetching the encap config file
stat() returns an empty  string.

The message "skip encap print env mtime: the container is not
running here" is printed by the agent, followed by a stack
caused by the code trying to parse this empty string instead
of bailing out.