Merge the collector CLI in the agent
The entry point is 'nodemgr collector cli'

If the user executes without root privilege, he has to provide
its collector user credentials or provide them through a config
file pointed by --config (or ~/.opensvc-cli by default).

If executed as root, the node credentials are used.

To achieve that behaviour, the check_privs() call in nodemgr.py is
now bypassed for "collector cli" and "print schedule". The
cmdworker.log is also not populated when the command is run non-
root.