Run stop*, unprovision and shutdown commands daemonized
The head process waits for the daemonized process to die to
exit.

Doing do, if the stop kills the head process to satisfy a
resource stop constraint, the daemonized process is not
affected and the action continues.

This scenario happens when the stop action is run from a shell
issued by a login through an IP handled by the service, or when
the shell CWD is under a fs resource mount point.

Interrupting the head process display a message stating the
action is still going on.