Align node and services environment parameters
The parameter describing the environment for the node was named
'host_mode', whereas for services it was named 'service_type'.

This was confusing and none of the keywords were clearly hinting
about 'environments' (as in DEV, PRD, ...).

The host_mode and service_type keywords are now deprecated (still
supported to ease agent upgrades), and both replaced by the 'env'
keyword.