Output formatting enhancements
* Deprecate the "json <command>". The preferred way is to specify
the --format=json parameter to the "print <command>" action.
The "json <command>" are no longer advertized in the command
help, but are still working for compatibily.

* Add a --format=csv formatter

* Make --cluster with a print command and --format=json aggregate
the data from remote nodes. The resulting dataset is a dict with
nodenames as top-level keys

* Make svcmgr print commands on multiple services and --format=json
aggregate the services data. The resulting dataset is a dict with
svcnames as top-level keys

* svcmgr print commands on a multiple services and non-json
format still print the results as concatenated results.

* Colorize itemized (default) format output

* Fix color ansi codes appearing in json status output

* Remove the --table parameter

* Remove json commands from bash completion