Remove the "parent process cwd is held by the %s resource" fs pre_action
This sanity check was meant to avoid kill the calling shell if its
cwd was in a service fs resource mountpoint, upon service stop.

Now the agent runs the stop action in a detached process, so the
stop action continues even after the calling shell is killed. So
the sanity check is no longer useful.