Fix wrong rtype reported to the collector for encap resources
Add resource type to the status_quad() returned data, hence
embed it in "json status" output and use it when preparing the
payload to send to the collector.

If the encap agent is not upgraded to embed the resource type in
"json status", fallback to an empty string, which is better than
a wrong information.

Evidently, it is better upgrade both master and encap agents.