Allow scoped commands to apply to specified resources
An action applied to resources specified by --rid <family|rid>...,
--subset, or --tags would not apply to zoned or dockerized
resources if the zone or docker resource is not one of the
selected resources.

This behaviour does not permit to stop zoned or dockerized
resources without stopping the container, which is not fair.

This patch inhibits the xtag logic in Resource::action() when
the command is scoped.