Degrade overall status if a res has warn or err status log messages
This patch is necessary to enable monitoring on resources. Up to
now, non-optimal resources were put in state "warn", triggering
a failover if the resource is monitored.

Better to keep the resource state aligned to the real availability
state, and have the presence of warnings cascade to the service
overall state.

We might introduce later a new resource parameter, instructing
the resource monitor to TOC when a resource has warnings or
errors.

Update the disk.md driver so that the resource state keeps up
when the resource is degraded.

Example:

testmd
overall                   warn
|- avail                  up
|  |- ip#0           .... up         128.1.11.1@lo
|  |- disk#00        .... up         loop /opt/testmd/dd1
|  |- disk#01        .... up         loop /opt/testmd/dd2
|  |- disk#02        .... up         loop /opt/testmd/dd3
|  |- disk#03        .... up         loop /opt/testmd/dd4
|  |- disk#04        .... up         loop /opt/testmd/dd5
|  |- disk#05        .... up         loop /opt/testmd/dd6
|  |- disk#10        .... up         md b742617b:d8a76908:5ec4db5e:5e4d920b
|  |                                 warn: clean, degraded
|  |