Skip syncrcopy if the rcg is in Periodic Mode in sync.hp3par
If the rcg is in Periodic Mode, the agent is not required to
ask for a resync through the syncrcopy command, as the array
is responsible for the periodic resync.

The resource driver still monitors if the resync is not happening
in the sync_max_delay threshold, and is still responsible for
the replication invertion when needed.