Report container hostname instead of name in the status payload
The collector tries to find if a received service status payload
is coming from the encap part of a service, using the vmname
field content as a nodename.

The agent encoded the container name in the vmname field instead
of the container hostname, so the collector failed to detect
the emitter is encap when container name!=hostname. Which lead
to undue entries in the svcmon table.