Wait for dockerd stopped
During service unprovision, the private dockerd is stopped and
the fs umounted. If the agent does not wait for dockerd shutdown,
the fs stop action will force-kill the daemon and produce
warnings.

Add a 10s wait loop, and send a kill sig if the term sig did not
work.

Also log the pid of the dockerd we signal.