Make sure OSVC_SERVICE_LINK is unset when using svcmgr directly
Zones and lxc remote commands can preserve the caller's env.
So if a "mysvc push" command is executed on the master agent,
a "svcmgr -s mysvc ..." command is triggered in the slave agents
with OSVC_SERVICE_LINK set, and the parser is loaded without
the -s option, thus outs the help message and exits.

This patch make sure we unset OSVC_SERVICE_LINK in the opensvc
shell wrapper when called directly.