Disable the service scheduler until /var/run/opensvc.<svname> exists
It is counterproductive to let the scheduler run until the
opensvc launcher has booted a service, because

1/ tasks waste cpu/ram/io resources at a moment we want them
available to start the services
2/ it causes lock contention between the boot action and the
resource_monitor or push_service_status tasks

This patch bypass all scheduled tasks until a per-service flag
exists in /var/run/.

This flag is created after any non-scheduled task is terminated.
The boot action from the startup script creates this flag, for
example.