Add a generic, collector ipam based, ip provisioner
The collector can describe networks and network segments.
A network segment:
* can be typed static, dynamic or reserved
* has responsibles

The collector also has a primary dns implementation.

A ip is considered free if not in a segment and if no dns
name is pointing to this ip.

The Rest API exposes all IPAM and DNS ops.

This ip provisionner logs to the collector with service credentials
and asks for a free ip in the specified network (resource network
option) or if not specified the network of the base ip on "ipdev".

The collector creates the dns record upon allocation, using the
same naming rules as the "dns update" codepath.

Example:

$ sudo testmd stopip --rid ip#1
11:45:09,597 ip#1        INFO    ifconfig lxcbr0:1 down
11:45:09,608 ip#1        INFO    checking 10.0.3.2 availability

$ sudo testmd unset --param ip#1.ipname
$ sudo testmd set --param ip#1.dns_name_suffix --value backend

$ sudo testmd provision --rid ip#1
11:46:01,016 ip#1        INFO    ip 10.0.3.5 allocated
11:46:01,023 ip#1        INFO    checking 10.0.3.5 availability
11:46:04,030 ip#1        INFO    ifconfig lxcbr0:1 10.0.3.5 netmask 255.255.255.0 up
11:46:04,047 ip#1        INFO    arping -U -c 1 -I lxcbr0 -s 10.0.3.5 10.0.3.5
11:46:04,382 ip#1        INFO    dns updated
11:46:04,496             INFO    send /opt/opensvc/etc/testmd.conf to collector
11:46:04,497             INFO    update /opt/opensvc/var/testmd/last_pushed_config timestamp

$ host testmd-backend.domnet.opensvc 10.0.3.11
Using domain server:
Name: 10.0.3.11
Address: 10.0.3.11#53
Aliases:

testmd-backend.domnet.opensvc has address 10.0.3.5