Move the logfile out of rcEnv
The logfile is only need upon logger configuration, so no need
to keep it around in rcEnv.

Its formatting is also moved out of the builder into to rcLogger.

The "logs" nodemgr and svcmgr action now go get the logfile
location directly into the log file handler.

With this patch, the unit tests can pop a Svc(svname) object.

Also add test units for the "logs" nodemgr and svcmgr action.