Command line options passing changes
Unit tests writers want to be able to instanciate a service and
submit multiple actions.

Using svcmgr.main(), the service is reinstanciated upon each call.

This patch makes the action entry point Svc::action() take the
full options dict as argument. It thus moves options mangling
logic from svcmgr to svc, where it belongs, and is a first step
towards considering Svc::options as a defaults storage and not
a storage for action options.

Also add a new test class test_svc_fs, which uses these new
capabilities, and can be used as a model for svc testing:

* class based for Svc() object persistence across multiple tests
* setup/teardown
* multiple use of Svc::action(action, options)