Move pull action code from node to svc
Teach svcBuilder::build_services() to create a dummy Svc()
instance if the create_instance kwarg is set to True.

Use this facility in the "pull" action codepath, so that the
pull method of the dummy service instance can be called without
special-casing in svcmgr.

Add a pull test to the suite.