Support a new rule data structure in the keyval comp object
This object is still compatible with the previous format.
The new format shifts the list of key/op/val dict into a top
level dict key named keys. At the same level than "keys", the
designer can specify a "path" key to point the object to the
file where the keys should be checked/fixed.

This new format does no longer require the file path to be
passed to the comp object as an argument.

Example legacy format:

[
  {
    "key": "AuthorizedKeysFile2",
    "value": " ",
    "op": "unset"
  },
  {
    "key": "AuthorizedKeysFile",
    "value": " ",
    "op": "unset"
  },
  {
    "key": "X11Forwarding",
    "value": "yes",
    "op": "="
  },
  {
    "key": "Banner",
    "value": "/etc/issue.net",
    "op": "="
  },
  {
    "key": "Banner",
    "value": "/etc/issue",
    "op": "unset"
  }
]

Example new format:

{
  "keys": [
    {
      "value": "bb",
      "key": "aa",
      "op": "="
    },
    {
      "value": "cc",
      "key": "bb",
      "op": "="
    }
  ],
  "path": "/tmp/test_keyval_with_fpath"
}

Example check run on both data format:

LOG:
aa=bb on target
bb=cc on target
AuthorizedKeysFile2 is not set, on target
AuthorizedKeysFile is not set, on target
X11Forwarding=yes on target
Banner=/etc/issue.net on target
Banner is not set to value /etc/issue, on target
aa=bb on target
bb=cc on target